move AbstractContainerAwareChekcerTestCase under ECS
